This is similar to snake and rock based game, written in x86 assembly language developed in DoxBox-X.
The rocks are randomly placed on the bottom row of the screen, and the chonker is drawn near the middle of the screen. The entire screen is scrolled upwards, so the rocks move upwards and the chonker leaves a tunnel behind itself. The chonker is controlled with the “a” key to move left and the “s” key to move right. The chonker must avoid the rocks for as long as possible. The “q” key quits the game.